Completed file upload and save, started on creating form for NFT
